Add specialization support and clean up
Removes the Identity resolver in favor of giving the unit struct its
functionality. Removes the rel_ptr! macro in favor of re-exporting
offset_of from memoffset. Fixes some old const_generics code and adds
specialization support for slices. Also allows slices with the identity
resolver to be archived in no_std. Adds the ArchiveBuffer struct and
switches tests to use that instead of ArchiveWriter so that they can be
run in no-std.