Fix cached vtable implementation, reduce size
Switch from a cell-based vtable cache implementation to an atomic-based
one. When using the vtable cache, the lowest significant bit of the cached
value will be used to determine if the vtable is cached or not. Because
of this, all impl ids are now OR'd with 1 to guarantee that the least
significant bit is always 1. The type registry additionally asserts that
the vtable pointers added to it do not have the least significant bit
set. If vtable caching is not turned on, then it falls back to using a
simple u64.