Separate TypeName into new crate, fix UB crash
Clean up crate TOMLs so that dependencies work correctly locally and
from crates.io. Separate TypeName functionality into a new type_name
crate and add primitive and std type implementations. Fix an issue with
the vtable macro: it used to make the data reference out of a null
pointer but that causes a crash when running in release. Instead, it now
creates the data pointer from a local MaybeUninit. The crash no longer
manifests.