Add macro arguments to derive and dyn macros
Adds an archived argument to the Archive derive that exposes the
archived type as the given name. This is mostly so that you can
implement stuff on it without the compiler complaining that generic
parameters aren't bound. Also adds a dyn_archive argument to rename the
trait + Archived trait. Add a few more trait implementations for core
and std types. Minor cleanup. Updates a few tests to test the new
features.