impl Archive, Deserialize, and Serialize for AlignedBytes and AlignedVec
The implementation strategy is to derive all of these trait impls. This
was straightforward for AlignedBytes since it is just POD, but it
required a little more effort for AlignedVec. Thankfully, by changing
the representation of AlignedVec to use an internal
Vec<AlignedBytes<16>>, most of the code got simpler. All of the manual
allocations went away, and much of the unsafe code went with it.

This does include a change to some of the test's expectations about what
should be returned by AlignedVec::capacity. Now, capacity is always a
multiple of ALIGNMENT=16. It is unlikely that anyone relied upon the
exact capacity. I think so long as reserve and reserve_exact follow the
same contract of reserving "enough" space for the additional bytes, we
do not need to call this a breaking change.

ArchivedAlignedVec also got some manual trait impls, similar to what
ArchivedVec has, and those impls were mostly copied from those of
ArchivedVec.