fixes const_generic_char test
Fixed by formatting the const generic value with Debug in stead of
Display.

Formatting with Debug creates the expected quotes in opposition to
Display.

This implementation should be sound for rust 1.68 but this code may
break in subtle or less subtle ways when new const-generic features
are added to the compiler. Added comments about this in the code.

Related issue #383