Issue 374 (#385)
* Added integration tests to rkyv_typename

This is primarily done for testing issue #374.

However, in addition to the original issue some other errors in the
generated code from the TypeName derive macro were encountered:
1) const generics suffer from essentially the same issue
2) A name collision in the generic parameter F which occurs in the
   build_type_name<F: FnMut(&str)>(...) method. The same name F is used
   in the generated code. In stead it should be given a name which is
   unlikely to clash with code that invokes the macro.

In addition to failing tests some trivial tests were added which do not
fail:

- basic_enum
- generic_one_param
- generic_with_trait_bounds
- generic_two_params
- basic_struct
- trivial_enum
- trivial_struct
- trivial_tuple_struct
- trivial_unit_struct

The following test fails to compile due to the name collision with the
generic parameter F:

- generic_name_colisions

The following tests fail to compile due to the original bug in
issue #374:

- lifetimes
- lifetimes_two_lifetimes_with_bounds
- combined_generic_type_and_lifetime

The following test fails to compile due to the analogous related
const-generic issue.

- const_generics

The following test checks checks some more potential errors and fails
to compile due to combinations of these issues.

- composite_cases

* Fix failing integration tests for rkyv_typename

Fixes issue #374 as well as #384 and adds a possible solution for
issue #383.

- Issue #384 is a minor fix so I took the liberty of fixing it here.
- Issue #383 is non orthogonal to #374. I.e. some code involving
issue #383 is necessary to make everything work. It may however
require further discussion so issue #383 should probably remain open.

* additional test cases for const generics

- one of these fails because the const char generic creates output
which does not surround the char with the expected quotes.

Also: cosmetics

* fixes const_generic_char test

Fixed by formatting the const generic value with Debug in stead of
Display.

Formatting with Debug creates the expected quotes in opposition to
Display.

This implementation should be sound for rust 1.68 but this code may
break in subtle or less subtle ways when new const-generic features
are added to the compiler. Added comments about this in the code.

Related issue #383