bug: fix validation errors with pointer values (#466)
I've been hitting an issue where validation fails (and I don't know
why), and I've been getting this error:

    error: Reading from cache archive failed: context error: unaligned
    pointer: ptr 0x7fff559663b8 unaligned for alignment 4

Which is quite confusing because `0x7fff559663b8 % 4 == 0`. What was
happening instead was that the address of the pointer was being printed
instead of the pointer itself.

There are a few ways to fix this, but I think this is a case where match
egonomics is harming things. That is, we match on `*self` and
specifically insert borrows where necessary. That way, everything else
is guaranteed to be used by value. Now my error message (which I'm still
trying to track down) is this:

    error: Reading from cache archive failed: context error: unaligned
    pointer: ptr 0x7f3f4556d585 unaligned for alignment 4

I wish it could be more useful (like mentioning the trace to the
data type it is tripping on, although that seems hard to do without
sacrificing something important), but at least the message itself is
correct now.